requirements.txt
flask
flask-cors
boto3
face_recognition
pillow
gunicorn
redis
rq
numpy
psutil